Solution to Calculate and Print the Area of a Circle

+++ Calculate and Print the Area of a Circle@@6788d5012db25efec946f943/code.c
@@ -1,8 +1,8 @@
 #include <stdio.h>
 
 int main() {
-    int a;
-    scanf("%d",&a);
-    printf("Area: %d",3.14*a*a);
+    float a;
+    scanf("%f",&a);
+    printf("Area: %.2f",3.14*a*a);
     return 0;
 }
